requirements.txt
# Standard libraries
numpy
pandas
matplotlib
# plotly
time
datetime
# scipy
# seaborn
# Data scrapping
requests
json
# Paths in notebooks
sys
os
# Data fetching
forex_python.converter
cryptocompare
yfinance
# GUI
tkinter
pyinstaller